code/requirements.txt
alembic==1.0.0
appdirs==1.4.3
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
blinker==1.4
cached-property==1.5.1
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
click==6.7
defusedxml==0.6.0
enum-compat==0.0.2
enum34==1.1.6
et-xmlfile==1.0.1
eventlet==0.22.0
Flask==1.1.1
Flask-Cors==3.0.3
Flask-Migrate==2.2.1
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.2
geopy==1.11.0
greenlet==0.4.13
gunicorn==19.8.1
httplib2==0.10.3
idna==2.8
isodate==0.6.0
isort==4.3.20
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10.1
lazy-object-proxy==1.4.1
lxml==4.4.1
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==3.0.0b16
mccabe==0.6.1
more-itertools==6.0.0
oauth2client==4.1.2
openpyxl==2.4.8
passlib==1.7.1
phonenumbers==8.9.3
Pillow==7.0.0
pluggy==0.9.0
polyline==1.3.2
psycopg2==2.8.5
py==1.8.0
pyasn1==0.3.3
pyasn1-modules==0.1.1
pycparser==2.18
PyJWT==1.5.2
pylint==2.3.1
pytest==4.3.0
python-dateutil==2.7.3
python-decouple==3.1
python-editor==1.0.3
python-engineio==1.7.0
python-socketio==1.8.1
pytz==2020.1
raven==6.10.0
redis==2.10.6
requests==2.22.0
requests-toolbelt==0.9.1
rsa==3.4.2
six==1.12.0
SQLAlchemy==1.2.11
typed-ast==1.4.1
urllib3==1.25.3
validate-email==1.3
Werkzeug==0.15.6
wrapt==1.11.1
XlsxWriter==1.1.8
zeep==3.4.0
